% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioFig.R
\name{plot_2DPCA}
\alias{plot_2DPCA}
\title{Plot 2D PCA}
\usage{
plot_2DPCA(
  expression,
  group,
  samplenames,
  scl,
  colors = NULL,
  shape = NULL,
  ggrepelLab = TRUE,
  size_gglab = 5,
  size_title = 1,
  point.size = 4,
  MahalanobisEllips = F,
  LegendName_Color = "group",
  LegendName_Shape = "shape",
  ...
)
}
\arguments{
\item{expression}{The expression table to use to create the PCA}

\item{group}{Factor specifying the groups for the samples}

\item{samplenames}{Vector specifying the names of the samples}

\item{scl}{logical. If TRUE the data is scaled}

\item{colors}{Optional vector of colors for the factor levels}

\item{shape}{Optional vector of shapes for the factor levels}

\item{ggrepelLab}{logical. If TRUE sample names are included as ggrepel labels}

\item{size_gglab}{Size of the ggrepel labels}

\item{size_title}{Size of plot title}

\item{point.size}{Size of the dots}

\item{MahalanobisEllips}{logical. If TRUE mahalanobis ellipses are included in the plot}

\item{LegendName_Color}{Title for the color legend}

\item{LegendName_Shape}{Title for the shape legend}
}
\value{
This function return a PCA plot
}
\description{
This function creates a Principal Component Analysis (PCA) plot
}
