% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioFig.R
\name{plot_ExploratoryPlots}
\alias{plot_ExploratoryPlots}
\title{Plot exploratory plots}
\usage{
plot_ExploratoryPlots(
  expression,
  group,
  samplenames,
  plottype = c("all", "PCA", "MDS", "Distance", "Expression"),
  transform = "no",
  colors = NULL,
  shape = NULL,
  LegendName_Color = "group",
  LegendName_Shape = "shape",
  ggrepelLab = TRUE,
  size_gglab = 5,
  size_title = 1,
  point.size = 4,
  scale = T,
  ntop = NULL,
  MahalanobisEllips = F,
  savePlots = TRUE,
  ...
)
}
\arguments{
\item{expression}{The expression table to use}

\item{group}{Factor specifying the groups for the samples or the name of the column containing the factor variable if the input is a deseq2 object}

\item{samplenames}{Vector specifying the names of the samples or the name of the column containing the sample names if the input is a deseq2 object}

\item{plottype}{Type of plot to generate, options are 'PCA','MDS','Distance','Expression' and 'all'}

\item{transform}{Apply a transformation to the data, available transformation are 'vst' and 'rlog'}

\item{colors}{Optional vector of colors for the factor levels}

\item{shape}{Optional vector of shapes for the factor levels}

\item{LegendName_Color}{Title for the color legend}

\item{LegendName_Shape}{Title for the shape legend (for PCA)}

\item{ggrepelLab}{logical. If TRUE sample names are included as ggrepel labels}

\item{size_gglab}{Size of the ggrepel labels}

\item{size_title}{Size of plots titles}

\item{point.size}{Size of the dots for the PCA and MDS plots}

\item{scale}{logical. If explicitly set to TRUE, the data is scaled for the PCA plot even if a transformation was applied}

\item{ntop}{Use n features that explain most of the variance}

\item{MahalanobisEllips}{logical. If TRUE mahalanobis ellipses are included in the PCA plot}

\item{savePlots}{logical, If TRUE it saves the plots automatically as pdf and svg}
}
\value{
This function returns one or four plots for data exploration. The available plots are Principal Component Analysis (PCA), Multi Dimensional Scaling (MDS), Heatmap of sample distances, and a Heatmap of the expression of all features.
}
\description{
This function is thought as a help to explore biological data. It can create four exploratory plots. PCA and MDS plots on the first row and Expression and Sample Distance heatmaps on the second row.
Individual plots can also be obtained
}
\examples{
set.seed(78)
m  <- matrix(rnorm(100) , nrow = 20)
rownames(m) <-  paste0(rep("gene",20))
colnames(m) <- c("S1","S2","S3", "S4","S5")
group <- c("T1","T1","T1","T2","T2")
plot_ExploratoryPlots(expression=m,group=group,samplenames=colnames(m),plottype="all",
LegendName_Color="Treatment",point.size=6)
}
