% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioFig.R
\name{plot_ExpressionHeatmap}
\alias{plot_ExpressionHeatmap}
\title{Expression Heatmap}
\usage{
plot_ExpressionHeatmap(
  expTable,
  Title = "Heatmap",
  sampleDGroup,
  clColumns = FALSE,
  FontSRow = 10,
  FontSColumn = 10,
  FontSTitle = 16,
  setseed = NULL,
  ...
)
}
\arguments{
\item{expTable}{The expression table to use to create the heatmap}

\item{Title}{Title of the plot}

\item{sampleDGroup}{A data frame specifying groups/factors for every sample}

\item{clColumns}{logical. If TRUE columns will be clustered}

\item{FontSRow}{Font size for the row labels}

\item{FontSColumn}{Font size for the column labels}

\item{FontSTitle}{Font size for the title}

\item{setseed}{Set seed to obtain the same colors}
}
\value{
It returns a heatmap of the expression of all features
}
\description{
This function creates an expression heatmap of all samples
}
