% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioFig.R
\name{plot_MDS}
\alias{plot_MDS}
\title{Multidimensional Scaling}
\usage{
plot_MDS(
  expression,
  group,
  point.size,
  LegendName_Color = "group",
  LegendName_Shape = "shape",
  ...
)
}
\arguments{
\item{expression}{Numerical matrix with samples as columns (e.g gene expression)}

\item{group}{Factor specifying the groups for the samples}

\item{point.size}{Size of the dots for the plot}

\item{LegendName_Color}{Title for the color legend}

\item{LegendName_Shape}{Title for the shape legend}
}
\value{
It returns a MDS plot
}
\description{
This function create a multidimensional scaling (MDS) plot that visualizing the level of (dis)similarity between individual cases of a dataset.
MDS is a distance-preserving algorithms defined by the pairwise distances of data points.
MDS takes a matrix D where Dij represents the dissimilarity between points i and j and produces a mapping on a lower dimension, preserving the dissimilarities as closely as possible.
There are many different ways of calculating dissimilarity among samples, Euclidean distance is used as default
}
\examples{
set.seed(78)
m  <- matrix(rnorm(100) , nrow = 20)
rownames(m) <-  paste0(rep("gene",20))
colnames(m) <- c("S1","S2","S3", "S4","S5")
group <- colnames(m)
plot_MDS(m, point.size = 5, group = group)
}
