% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioFig.R
\name{plot_SampleDistance}
\alias{plot_SampleDistance}
\title{Sample distance heatmap}
\usage{
plot_SampleDistance(expression, ...)
}
\arguments{
\item{expression}{Numerical matrix with samples as columns (e.g gene expression)}
}
\value{
It returns a heatmap plot showing  dissimilarity/similarity between the samples
}
\description{
This function is created to show variable dissimilarity/similarity in a heatmap using distance or correlation as method.
}
